+++ shuup_tests/admin/test_urls.py
@@ -13,6 +13,7 @@ from django.core.exceptions import ImproperlyConfigured
 from shuup.admin.utils.urls import admin_url, get_model_url, NoModelUrl
 from shuup.core.models import Product
 from shuup_tests.admin.utils import admin_only_urls
+from shuup_tests.utils.faux_users import StaffUser
 
 
 def test_model_url():
@@ -22,6 +23,30 @@ def test_model_url():
         p = Product()
         p.pk = 3
         assert get_model_url(p)
+
+
+def test_model_url_with_permissions():
+    permissions = set(["shuup.add_product", "shuup.delete_product", "shuup.change_product"])
+    p = Product()
+    p.pk = 3
+
+    # If no user is given, don't check for permissions
+    assert get_model_url(p)
+
+    # If a user is given and no permissions are provided, check for default model permissions
+    user = StaffUser()
+    with pytest.raises(NoModelUrl):
+        assert get_model_url(p, user=user)
+
+    # If a user is given and permissions are provided, check for those permissions
+    assert get_model_url(p, user=user, required_permissions=())
+    with pytest.raises(NoModelUrl):
+        assert get_model_url(p, user=user, required_permissions=["shuup.add_product"])
+
+    # Confirm that url is returned with correct permissions
+    user.permissions = permissions
+    assert get_model_url(p, user=user)
+    assert get_model_url(p, user=user, required_permissions=permissions)
 
 
 def test_invalid_admin_url():